% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monomediate.R
\name{monomediate}
\alias{monomediate}
\title{One-step mediation estimator under monotonicity assumption}
\usage{
monomediate(data, npsem, learners, folds, y_family)
}
\arguments{
\item{data}{}

\item{npsem}{}

\item{learners}{}

\item{folds}{}

\item{y_family}{}
}
\description{
One-step mediation estimator under monotonicity assumption
}
