% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monomediate_gcomp.R
\name{monomediate_gcomp}
\alias{monomediate_gcomp}
\title{Parametric mediation estimator under monotonicity assumption}
\usage{
monomediate_gcomp(data, npsem, y_form, z_form, m_form, y_family)
}
\arguments{
\item{data}{}

\item{npsem}{}

\item{y_form}{}

\item{z_form}{}

\item{m_form}{}

\item{y_family}{}
}
\description{
Parametric mediation estimator under monotonicity assumption
}
